refactor(scripts): modularize verify-handoff-lead-to-plan.js into verify-l2p/ modules
SD-LEO-REFACTOR-VERIFY-L2P-001 - Refactors 1512 LOC file into 7 focused modules:
- constants.js (113 LOC) - SD requirements, type patterns, validation config
- sd-validation.js (286 LOC) - SD completeness validation logic
- prd-readiness.js (397 LOC) - 6 PRD readiness validation methods
- type-detection.js (102 LOC) - Auto-detect SD type based on keywords
- environment.js (80 LOC) - Environment readiness checks
- handoff-execution.js (211 LOC) - Create handoff records and rejection handling
- index.js (544 LOC) - LeadToPlanVerifier class orchestration + CLI

Original file now re-exports from modular structure for backward compatibility.
All modules pass syntax validation. Supports independent testing of validation domains.
